Basic fix for issue #23
Adding yield() enables the ESP8266 to do other things while waiting and prevents reboots due to watchdog timer timeouts..

I had a hard tme to find out that the while-loop was the cause for the reboots....

That´s exactly what yield() is for -> Passes control to other tasks when called. Ideally yield() should be used in functions that will take awhile to complete.